Update to migrate domain models to suitable dir and namespace
Also, updates all API consumer refs to these classes
