+++ Models/Large.cs
@@ -1,17 +1,17 @@
 ﻿
 using System.Text;
 using System.IO;
-using SailSafe.Classes;
+using SailSafe.Models;
 
-namespace SailSafe
+namespace SailSafe.Models
 {
-    class Van : Vehicle
+    class Large : Vehicle
     {
-        public override int Length { get { return 5; } }
+        public override int Length { get { return 15; } }
 
-        public Van()
+        public Large()
         {
-            this.Type = VehicleType.Van;
+            this.Type = VehicleType.Large;
         }
 
         ///// <summary>
@@ -20,19 +20,19 @@ namespace SailSafe
         ///// <param name="artist">The name of the artist</param>
         ///// <param name="title">The name of the work</param>
         ///// <param name="price">The retail price (in pounds sterling)</param>
-        //public Van(string name, string license, int time, bool directionNtoS, bool directionStoN)
+        //public Large(string name, string license, int time, bool directionNtoS, bool directionStoN)
         //    : base(name, license, time, directionNtoS, directionStoN)
         //{
-        //    this.Type = VehicleType.Van;
+        //    this.Type = VehicleType.Large;
         //}
 
         ///// <summary>
         ///// Constructor for loading the data from a text string
         ///// </summary>
         ///// <param name="source"> ',' delimited string of data read back in from a file </param>
-        //public Van(string source)
+        //public Large(string source)
         //{
-        //    this.Type = VehicleType.Van;
+        //    this.Type = VehicleType.Large;
 
         //    string[] data = source.Split(',');
         //    Builder(data[1], data[2], int.Parse(data[3]), bool.Parse(data[4]),
