+++ SailSafe.csproj
@@ -61,17 +61,17 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="Classes\Booking.cs" />
-    <Compile Include="Classes\SailingDirection.cs" />
+    <Compile Include="Models\Booking.cs" />
+    <Compile Include="Models\SailingDirection.cs" />
     <Compile Include="Forms\CancelBooking.cs">
       <SubType>Form</SubType>
     </Compile>
     <Compile Include="Forms\CancelBooking.Designer.cs">
       <DependentUpon>CancelBooking.cs</DependentUpon>
     </Compile>
-    <Compile Include="Classes\Car.cs" />
-    <Compile Include="Classes\Lane.cs" />
-    <Compile Include="Classes\Large.cs" />
+    <Compile Include="Models\Car.cs" />
+    <Compile Include="Models\Lane.cs" />
+    <Compile Include="Models\Large.cs" />
     <Compile Include="Forms\LocateBooking.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -90,7 +90,7 @@
     <Compile Include="Forms\NewBookingForm.Designer.cs">
       <DependentUpon>NewBookingForm.cs</DependentUpon>
     </Compile>
-    <Compile Include="Classes\Sailing.cs" />
+    <Compile Include="Models\Sailing.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
     <Compile Include="Forms\SignIn.cs">
       <SubType>Form</SubType>
@@ -98,8 +98,8 @@
     <Compile Include="Forms\SignIn.Designer.cs">
       <DependentUpon>SignIn.cs</DependentUpon>
     </Compile>
-    <Compile Include="Classes\Van.cs" />
-    <Compile Include="Classes\Vehicle.cs" />
+    <Compile Include="Models\Van.cs" />
+    <Compile Include="Models\Vehicle.cs" />
     <Compile Include="Forms\ViewSailings.cs">
       <SubType>Form</SubType>
     </Compile>
@@ -144,7 +144,7 @@
       <DesignTimeSharedInput>True</DesignTimeSharedInput>
     </Compile>
     <Compile Include="Interfaces\IRepository.cs" />
-    <Compile Include="Classes\VehicleType.cs" />
+    <Compile Include="Models\VehicleType.cs" />
     <Compile Include="Intermediaries\BookingManager.cs" />
     <Compile Include="Intermediaries\BookingRepository.cs" />
     <Compile Include="Utilities\DateTimeHelper.cs" />
